refactor: Introduces `linear-witness-test-utils`
