+++ linear-witness-test-utils/README.md
@@ -1,0 +1,3 @@
+# linear-witness-test
+
+This package houses the test utilities for `linear-witness` and tests for witness-based allocations provided there.
