+++ linear-witness-test-utils/linear-witness-test-utils.cabal
@@ -1,0 +1,69 @@
+cabal-version: 1.12
+
+-- This file has been generated from package.yaml by hpack version 0.35.2.
+--
+-- see: https://github.com/sol/hpack
+
+name:           linear-witness-test-utils
+version:        0.1.0.0
+description:    Please see the README on GitHub at <https://github.com/konn/linear-witness-test#readme>
+homepage:       https://github.com/konn/linear-witness-test#readme
+bug-reports:    https://github.com/konn/linear-witness-test/issues
+author:         Hiromi ISHII
+maintainer:     konn.jinro_at_gmail.com
+copyright:      2023 (c) Hiromi ISHII
+license:        BSD3
+license-file:   LICENSE
+build-type:     Simple
+extra-source-files:
+    README.md
+    CHANGELOG.md
+
+source-repository head
+  type: git
+  location: https://github.com/konn/linear-witness-test
+
+library
+  exposed-modules:
+      Linear.Witness.Token.TestUtils
+  other-modules:
+      Paths_linear_witness_test_utils
+  hs-source-dirs:
+      src
+  ghc-options: -Wall -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wmissing-export-lists -Wmissing-home-modules -Wpartial-fields -Wredundant-constraints
+  build-depends:
+      base >=4.7 && <5
+    , containers
+    , dlist
+    , falsify
+    , foldl
+    , linear-base
+    , linear-witness
+    , tasty
+    , tasty-hunit
+    , vector
+  default-language: Haskell2010
+
+test-suite linear-witness-test
+  type: exitcode-stdio-1.0
+  main-is: Spec.hs
+  other-modules:
+      Paths_linear_witness_test_utils
+  hs-source-dirs:
+      test
+  ghc-options: -Wall -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wmissing-export-lists -Wmissing-home-modules -Wpartial-fields -Wredundant-constraints -threaded -rtsopts -with-rtsopts=-N
+  build-tool-depends:
+      tasty-discover:tasty-discover
+  build-depends:
+      base >=4.7 && <5
+    , containers
+    , dlist
+    , falsify
+    , foldl
+    , linear-base
+    , linear-witness
+    , linear-witness-test-utils
+    , tasty
+    , tasty-hunit
+    , vector
+  default-language: Haskell2010
