+++ linear-witness-test-utils/package.yaml
@@ -1,6 +1,6 @@
-name:                linear-array-extra
+name:                linear-witness-test-utils
 version:             0.1.0.0
-github:              "konn/linear-array-extra"
+github:              "konn/linear-witness-test"
 license:             BSD3
 author:              "Hiromi ISHII"
 maintainer:          "konn.jinro_at_gmail.com"
@@ -17,15 +17,20 @@ extra-source-files:
 # To avoid duplicated efforts in documentation and dealing with the
 # complications of embedding Haddock markup inside cabal files, it is
 # common to point users to the README.md file.
-description:         Please see the README on GitHub at <https://github.com/konn/linear-array-extra#readme>
+description:         Please see the README on GitHub at <https://github.com/konn/linear-witness-test#readme>
 
 dependencies:
 - base >= 4.7 && < 5
-- vector
-- linear-base >= 0.2
-- linear-generics
+- linear-base
 - linear-witness
-- primitive
+- containers
+- dlist
+- falsify
+- foldl
+- linear-base
+- tasty
+- tasty-hunit
+- vector
 
 ghc-options:
 - -Wall
@@ -40,26 +45,9 @@ ghc-options:
 
 library:
   source-dirs: src
-  c-sources:
-  - cbits/*.c
-  other-modules:
-  - Foreign.Atomic.Internal
-  dependencies:
-  - ghc-prim
-
-executables:
-  linear-array-extra-exe:
-    main:                Main.hs
-    source-dirs:         app
-    ghc-options:
-    - -threaded
-    - -rtsopts
-    - -with-rtsopts=-N
-    dependencies:
-    - linear-array-extra
 
 tests:
-  linear-array-extra-test:
+  linear-witness-test:
     main:                Spec.hs
     source-dirs:         test
     ghc-options:
@@ -69,13 +57,5 @@ tests:
     build-tools:
     - tasty-discover
     dependencies:
-    - linear-array-extra
+    - linear-witness
     - linear-witness-test-utils
-    - containers
-    - dlist
-    - falsify
-    - foldl
-    - linear-base
-    - tasty
-    - tasty-hunit
-    - vector
